Return and handle an error
greetings.go
Change the function so that it returns two values: a string and an error. Your caller will check the second value to see if an error ocurred. (Any Go function can return multiple values)
Import the Go standard library errors package so you can use its errors.
Add and if statement to check and invalid request (an empty string where the name should be) and return an error if the request is invalid. The errors.New function returns an error with your message inside.
Add nil (meaning no error) as a second value in the successful return. That way, the caller can see that function succeeded.
hello.go file
Configure the log package to print the command name (greetings:) at the start of its log messages, without a time stamp or source file information.
Assign both of the Hello return values, including the error, to variables.
Change the Hello argument from Glady's name to an empty string, so you can try out your error-handling code.
Look for a non-nil error value. There's no sense continuing in this case.
Used the functions in the standard library's log package to output error information. If you get an error, you use the log package's Fatal function to print the error and stop the program.